test/vhost: fix parted error
Vhost nightly test fails on Error: Partition have been written, but
we have been unable to inform the kernel of the change, probably
because it/they are in use.

+++ test/vhost/integrity/integrity_vm.sh
@@ -20,12 +20,17 @@ trap "exit 1" SIGINT SIGTERM EXIT
 
 for fs in $fs; do
 	for dev in $devs; do
+		i=0
 		parted_cmd="parted -s /dev/${dev}"
 
 		echo "INFO: Creating partition table on disk using: $parted_cmd mklabel gpt"
 		$parted_cmd mklabel gpt
+		if ! ($parted_cmd print | grep -q gpt); then
+			[ $i -lt 100 ] || break
+			i=$((i+1))
+			sleep 0.1
+		fi
 		$parted_cmd mkpart primary 2048s 100%
-		sleep 2
 
 		mkfs_cmd="mkfs.$fs"
 		if [[ $fs == "ntfs" ]]; then
@@ -33,8 +38,13 @@ for fs in $fs; do
 		fi
 		mkfs_cmd+=" /dev/${dev}1"
 		echo "INFO: Creating filesystem using: $mkfs_cmd"
-		wipefs -a /dev/${dev}1
 		$mkfs_cmd
+		while [ ! -e "/dev/${dev}1" ]; do
+			[ $i -lt 100 ] || break
+			i=$((i+1))
+			sleep 0.1
+		done
+		wipefs -a /dev/${dev}1
 
 		mkdir -p /mnt/${dev}dir
 		mount -o sync /dev/${dev}1 /mnt/${dev}dir
@@ -51,6 +61,7 @@ for fs in $fs; do
 	for dev in $devs; do
 		umount /mnt/${dev}dir
 		rm -rf /mnt/${dev}dir
+		parted -s /dev/${dev} rm 1
 
 		stats=( $(cat /sys/block/$dev/stat) )
 		echo ""
